+++ widgets/sysbusdev_description_settings.py
@@ -8,10 +8,10 @@ class SystemBusDeviceDescriptionSettingsWidget(DeviceDescriptionSettingsWidget):
     def __init__(self, *args, **kw):
         DeviceDescriptionSettingsWidget.__init__(self,
             [
-                ("out_irq_num", _("Output IRQ quantity"), long),
-                ("in_irq_num", _("Input IRQ quantity"), long),
-                ("mmio_num", _("MMIO quantity"), long),
-                ("pio_num", _("PMIO (PIO) quantity"), long)
+                ("out_irq_num", _("Output IRQ quantity"), int),
+                ("in_irq_num", _("Input IRQ quantity"), int),
+                ("mmio_num", _("MMIO quantity"), int),
+                ("pio_num", _("PMIO (PIO) quantity"), int)
             ],
             *args, **kw
         )
